Prediction-of-Graduate-Admissions
Based on the variations in each aspect, we can predict the admission chances for a certain student using the provided student information. A college accepts applications based on a variety of factors. As a result, we may acquire the results by employing the classification technique. In our model Classification has two results: true (1) or false (0), and the generated value is correct with any of these possibilities. It is determined by the type of classifiers or algorithms used.